requirements.txt
et-xmlfile==1.1.0
ImageHash==4.3.1
imageio==2.31.1
lazy_loader==0.3
networkx==3.1
numpy==1.25.2
opencv-python==4.8.0.76
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
scikit-image==0.21.0
scipy==1.11.2
six==1.16.0
tifffile==2023.8.12
tzdata==2023.3
